feat(nvim): switch from packer to lazy
https://github.com/folke/lazy.nvim
https://www.lazyvim.org
