--- .config/nvim/after/plugin/autopairs.lua
@@ -1,42 +1,0 @@
---[[
--- https://github.com/windwp/nvim-autopairs
---]]
-local status_ok, npairs = pcall(require, 'nvim-autopairs')
-if not status_ok then
-  print('autopairs is not installed')
-  return
-end
-
-npairs.setup({
-  disable_filetype = { 'TelescopePrompt', 'spectre_panel' },
-  disable_in_macro = true, -- disable when recording or executing a macro
-  disable_in_visualblock = true, -- disable when insert after visual block mode
-  ignored_next_char = string.gsub([[ [%w%%%'%[%"%.] ]], '%s+', ''),
-  enable_moveright = true,
-  enable_afterquote = true, -- add bracket pairs after quote
-  enable_check_bracket_line = true, --- check bracket in same line
-  map_bs = true, -- map the <BS> key
-  map_c_h = false, -- Map the <C-h> key to delete a pair
-  map_c_w = false, -- map <c-w> to delete a pair if possible
-  check_ts = true, -- enable with treesitter
-  ts_config = {
-    lua = { 'string', 'source' },
-    javascript = { 'string', 'template_string' },
-    java = false,
-  },
-  fast_wrap = {
-    map = '<M-e>',
-    chars = { '{', '[', '(', '"', "'" },
-    pattern = string.gsub([[ [%'%"%)%>%]%)%}%,] ]], '%s+', ''),
-    offset = 0, -- Offset from pattern match
-    end_key = '$',
-    keys = 'qwertyuiopzxcvbnmasdfghjkl',
-    check_comma = true,
-    highlight = 'PmenuSel',
-    highlight_grey = 'LineNr',
-  },
-})
-
-local cmp_autopairs = require('nvim-autopairs.completion.cmp')
-local cmp = require('cmp')
-cmp.event:on('confirm_done', cmp_autopairs.on_confirm_done({ map_char = { tex = '' } }))
