--- .config/nvim/after/plugin/colorizer.lua
@@ -1,13 +1,0 @@
---[[
--- https://github.com/norcalli/nvim-colorizer.lua
---]]
-local status_ok, colorizer = pcall(require, 'colorizer')
-if not status_ok then
-  print('colorizer is not installed')
-  return
-end
-
-colorizer.setup({
-  css = { css = true },
-  html = { names = false },
-})
