--- .config/nvim/after/plugin/mkdir.lua
@@ -1,8 +1,0 @@
---[[
--- https://github.com/jghauser/mkdir.nvim
---]]
-local status_ok, _ = pcall(require, 'mkdir')
-if not status_ok then
-  print('mkdir is not installed')
-  return
-end
