--- .config/nvim/after/plugin/telescope.lua
@@ -1,123 +1,0 @@
---[[
--- https://github.com/nvim-telescope/telescope.nvim
---]]
-local status_ok, telescope = pcall(require, 'telescope')
-if not status_ok then
-  print('telescope is not installed')
-  return
-end
-
-local actions = require('telescope.actions')
-local picker_height = 0.25
-local picker_width = 0.85
-
-telescope.setup({
-  defaults = {
-
-    prompt_prefix = '  ',
-    selection_caret = ' ',
-    path_display = { 'smart' },
-
-    mappings = {
-      i = {
-        ['<C-j>'] = actions.cycle_history_next,
-        ['<C-k>'] = actions.cycle_history_prev,
-
-        ['<C-n>'] = actions.move_selection_next,
-        ['<S-n>'] = actions.move_selection_previous,
-
-        ['<C-c>'] = actions.close,
-
-        ['<Down>'] = actions.move_selection_next,
-        ['<Up>'] = actions.move_selection_previous,
-
-        ['<CR>'] = actions.select_default,
-        ['<C-x>'] = actions.select_horizontal,
-        ['<C-v>'] = actions.select_vertical,
-        ['<C-t>'] = actions.select_tab,
-
-        ['<C-u>'] = actions.preview_scrolling_up,
-        ['<C-d>'] = actions.preview_scrolling_down,
-
-        ['<PageUp>'] = actions.results_scrolling_up,
-        ['<PageDown>'] = actions.results_scrolling_down,
-
-        ['<Tab>'] = actions.toggle_selection + actions.move_selection_worse,
-        ['<S-Tab>'] = actions.toggle_selection + actions.move_selection_better,
-        ['<C-q>'] = actions.send_to_qflist + actions.open_qflist,
-        ['<M-q>'] = actions.send_selected_to_qflist + actions.open_qflist,
-        ['<C-l>'] = actions.complete_tag,
-        ['<C-_>'] = actions.which_key, -- keys from pressing <C-/>
-      },
-
-      n = {
-        ['<esc>'] = actions.close,
-        ['<CR>'] = actions.select_default,
-        ['<C-x>'] = actions.select_horizontal,
-        ['<C-v>'] = actions.select_vertical,
-        ['<C-t>'] = actions.select_tab,
-
-        ['<Tab>'] = actions.toggle_selection + actions.move_selection_worse,
-        ['<S-Tab>'] = actions.toggle_selection + actions.move_selection_better,
-        ['<C-q>'] = actions.send_to_qflist + actions.open_qflist,
-        ['<M-q>'] = actions.send_selected_to_qflist + actions.open_qflist,
-
-        ['j'] = actions.move_selection_next,
-        ['k'] = actions.move_selection_previous,
-        ['H'] = actions.move_to_top,
-        ['M'] = actions.move_to_middle,
-        ['L'] = actions.move_to_bottom,
-
-        ['<Down>'] = actions.move_selection_next,
-        ['<Up>'] = actions.move_selection_previous,
-        ['gg'] = actions.move_to_top,
-        ['G'] = actions.move_to_bottom,
-
-        ['<C-u>'] = actions.preview_scrolling_up,
-        ['<C-d>'] = actions.preview_scrolling_down,
-
-        ['<PageUp>'] = actions.results_scrolling_up,
-        ['<PageDown>'] = actions.results_scrolling_down,
-
-        ['?'] = actions.which_key,
-      },
-    },
-  },
-  pickers = {
-    find_files = {
-      theme = 'dropdown',
-      previewer = false,
-      layout_config = {
-        height = picker_height,
-        width = picker_width,
-      },
-    },
-    git_files = {
-      theme = 'dropdown',
-      previewer = false,
-      layout_config = {
-        height = picker_height,
-        width = picker_width,
-      },
-    },
-    buffers = {
-      theme = 'dropdown',
-      previewer = false,
-      layout_config = {
-        height = picker_height,
-        width = picker_width,
-      },
-    },
-  },
-  extensions = {},
-})
-
-local builtin = require('telescope.builtin')
-vim.keymap.set('n', '<C-p>', builtin.find_files, {})
-vim.keymap.set('n', '<C-n>', builtin.git_files, {})
-vim.keymap.set('n', '<C-s>', builtin.buffers, {})
-vim.keymap.set('n', '<leader>fh', builtin.help_tags, {})
-vim.keymap.set('n', '<leader>ff', builtin.live_grep, {})
-vim.keymap.set('n', '<leader>ps', function()
-  builtin.grep_string({ search = vim.fn.input('Grep > ') })
-end)
