--- .config/nvim/lua/netrw.lua
@@ -1,7 +1,0 @@
---[[
--- Netrw, the vim explorer's configurations.
---]]
-
-vim.g.netrw_keepdir = 0 -- Avoid error when moving files
-vim.g.netrw_localcopydircmd = 'cp -r' -- Enable recursive copying
-vim.g.netrw_winsize = 30
