+++ .config/nvim/lua/plugins/lsp.lua
@@ -1,0 +1,287 @@
+return {
+  'VonHeikemen/lsp-zero.nvim', -- https://github.com/VonHeikemen/lsp-zero.nvim
+  dependencies = {
+    -- LSP Support
+    { 'neovim/nvim-lspconfig' },
+    { 'williamboman/mason.nvim' },
+    { 'williamboman/mason-lspconfig.nvim' },
+
+    -- Autocompletion
+    { 'hrsh7th/nvim-cmp' },
+    { 'hrsh7th/cmp-buffer' },
+    { 'hrsh7th/cmp-path' },
+    { 'saadparwaiz1/cmp_luasnip' },
+    { 'hrsh7th/cmp-nvim-lsp' },
+    { 'hrsh7th/cmp-nvim-lua' },
+
+    -- Snippets
+    { 'L3MON4D3/LuaSnip' },
+    { 'rafamadriz/friendly-snippets' },
+    { 'b0o/schemastore.nvim' },
+  },
+  config = function()
+    local lsp = require('lsp-zero')
+
+    lsp.set_preferences({
+      suggest_lsp_servers = true,
+      setup_servers_on_start = true,
+      set_lsp_keymaps = true,
+      configure_diagnostics = true,
+      cmp_capabilities = true,
+      manage_nvim_cmp = true,
+      call_servers = 'local',
+      sign_icons = {
+        error = '✘',
+        warn = '▲',
+        hint = '⚑',
+        info = '',
+      },
+    })
+
+    lsp.ensure_installed({
+      'bashls',
+      'diagnosticls',
+      'dockerls',
+      'gopls',
+      'html',
+      'jsonls',
+      'marksman',
+      'pyright',
+      'sqls',
+      'lua_ls',
+      'terraformls',
+      'vimls',
+      'yamlls',
+    })
+
+    -- Fix Undefined global 'vim'
+    lsp.configure('sumneko_lua', {
+      settings = {
+        Lua = {
+          diagnostics = {
+            globals = { 'vim' },
+          },
+        },
+      },
+    })
+
+    local cmp = require('cmp')
+    local cmp_select = { behavior = cmp.SelectBehavior.Select }
+    local cmp_mappings = lsp.defaults.cmp_mappings({
+      ['<C-p>'] = cmp.mapping.select_prev_item(cmp_select),
+      ['<C-n>'] = cmp.mapping.select_next_item(cmp_select),
+      ['<C-y>'] = cmp.mapping.confirm({ select = true }),
+      ['<C-Space>'] = cmp.mapping.complete(),
+    })
+
+    -- disable completion with tab
+    -- this helps with copilot setup
+    cmp_mappings['<Tab>'] = nil
+    cmp_mappings['<S-Tab>'] = nil
+
+    lsp.setup_nvim_cmp({
+      mapping = cmp_mappings,
+    })
+
+    lsp.set_preferences({
+      suggest_lsp_servers = false,
+      sign_icons = {
+        error = '',
+        warn = '',
+        hint = '',
+        info = '',
+      },
+    })
+
+    lsp.on_attach(function(client, bufnr)
+      local opts = { buffer = bufnr, remap = false }
+
+      if client.name == 'eslint' then
+        vim.cmd.LspStop('eslint')
+        return
+      end
+
+      vim.keymap.set('n', 'gD', vim.lsp.buf.declaration, opts)
+      vim.keymap.set('n', 'gd', vim.lsp.buf.definition, opts)
+      vim.keymap.set('n', 'gd', vim.lsp.buf.implementation, opts)
+      vim.keymap.set('n', 'gl', vim.diagnostic.open_float, opts)
+      vim.keymap.set('n', 'K', vim.lsp.buf.hover, opts)
+      vim.keymap.set('n', '<C-k>', vim.lsp.buf.signature_help, opts)
+      vim.keymap.set('n', '[d', vim.diagnostic.goto_next, opts)
+      vim.keymap.set('n', ']d', vim.diagnostic.goto_prev, opts)
+      vim.keymap.set('n', '<leader>ca', vim.lsp.buf.code_action, opts)
+      vim.keymap.set('n', '<leader>rn', vim.lsp.buf.rename, opts)
+    end)
+
+    lsp.setup()
+
+    vim.diagnostic.config({
+      virtual_text = true,
+    })
+
+    --[[
+-- https://github.com/hrsh7th/nvim-cmp
+-- https://github.com/L3MON4D3/LuaSnip
+--]]
+    local cmp = require('cmp')
+    local luasnip = require('luasnip')
+
+    local kind_icons = {
+      Text = '',
+      Method = 'm',
+      Function = '',
+      Constructor = '',
+      Field = '',
+      Variable = '',
+      Class = '',
+      Interface = '',
+      Module = '',
+      Property = '',
+      Unit = '',
+      Value = '',
+      Enum = '',
+      Keyword = '',
+      Snippet = '',
+      Color = '',
+      File = '',
+      Reference = '',
+      Folder = '',
+      EnumMember = '',
+      Constant = '',
+      Struct = '',
+      Event = '',
+      Operator = '',
+      TypeParameter = '',
+    } -- find more at: https://www.nerdfonts.com/cheat-sheet
+
+    local check_backspace = function()
+      local col = vim.fn.col('.') - 1
+      return col == 0 or vim.fn.getline('.'):sub(col, col):match('%s')
+    end
+
+    cmp.setup({
+      snippet = {
+        -- REQUIRED - a snippet engine
+        expand = function(args)
+          luasnip.lsp_expand(args.body)
+        end,
+      },
+      mapping = {
+        ['<C-b>'] = cmp.mapping(cmp.mapping.scroll_docs(-4), { 'i', 'c' }),
+        ['<C-f>'] = cmp.mapping(cmp.mapping.scroll_docs(4), { 'i', 'c' }),
+        ['<C-Space>'] = cmp.mapping(cmp.mapping.complete(), { 'i', 'c' }),
+        ['<C-y>'] = cmp.config.disable, -- Specify `cmp.config.disable` if you want to remove the default `<C-y>` mapping.
+        ['<C-e>'] = cmp.mapping({
+          i = cmp.mapping.abort(),
+          c = cmp.mapping.close(),
+        }),
+        -- Accept currently selected item.
+        -- `select` is set to `false` to only explicitly selected items on confirmation with Enter.
+        ['<CR>'] = cmp.mapping.confirm({ select = false }),
+        ['<Tab>'] = cmp.mapping(function(fallback)
+          if cmp.visible() then
+            cmp.select_next_item()
+          elseif luasnip.expandable() then
+            luasnip.expand()
+          elseif luasnip.expand_or_jumpable() then
+            luasnip.expand_or_jump()
+          elseif check_backspace() then
+            fallback()
+          else
+            fallback()
+          end
+        end, {
+          'i',
+          's',
+        }),
+        ['<S-Tab>'] = cmp.mapping(function(fallback)
+          if cmp.visible() then
+            cmp.select_prev_item()
+          elseif luasnip.jumpable(-1) then
+            luasnip.jump(-1)
+          else
+            fallback()
+          end
+        end, {
+          'i',
+          's',
+        }),
+      },
+      formatting = {
+        fields = { 'kind', 'abbr', 'menu' },
+        format = function(entry, vim_item)
+          -- Kind icons
+          vim_item.kind = string.format('%s', kind_icons[vim_item.kind])
+          -- vim_item.kind = string.format('%s %s', kind_icons[vim_item.kind], vim_item.kind) -- This concatonates the icons with the name of the item kind
+          vim_item.menu = ({
+            nvim_lsp = '[LSP]',
+            luasnip = '[Snippet]',
+            buffer = '[Buffer]',
+            path = '[Path]',
+          })[entry.source.name]
+          return vim_item
+        end,
+      },
+      sources = {
+        { name = 'nvim_lsp' },
+        { name = 'luasnip' },
+        { name = 'buffer' },
+        { name = 'path' },
+      },
+      confirm_opts = {
+        behavior = cmp.ConfirmBehavior.Replace,
+        select = false,
+      },
+      window = {
+        documentation = {
+          border = { '╭', '─', '╮', '│', '╯', '─', '╰', '│' },
+        },
+      },
+      experimental = {
+        ghost_text = false,
+        native_menu = false,
+      },
+    })
+
+    -- Setup Json SchemaStore via https://github.com/b0o/SchemaStore.nvim
+    require('lspconfig').jsonls.setup({
+      settings = {
+        json = {
+          validate = { enable = true },
+          schemas = require('schemastore').json.schemas({
+            -- https://github.com/SchemaStore/schemastore/blob/master/src/api/json/catalog.json
+            select = {
+              'Common types for all schemas',
+              -- "dependabot-v2.json",
+              'detekt.yml',
+              'dockerd.json',
+              'docker-compose.yml',
+              'Helm Chart.yaml',
+              'Helm Chart.lock',
+              'helmfile',
+              'GitHub Action',
+              'GitHub Workflow',
+              'GitHub Workflow Template Properties',
+              'Hadolint',
+              'JSON-API',
+              'kustomization.yaml',
+              'Liquibase',
+              'openapi.json',
+              'semantic-release',
+              'Starship',
+              'Swagger API 2.0',
+            },
+            replace = {
+              ['dependabot-v2.json'] = {
+                name = 'dependabot-v2.json',
+                description = "Override dependabot-v2.json, because it's only supported yml file ending",
+                fileMatch = { '**/.github/dependabot.yml', '**/.github/dependabot.yaml' },
+                url = 'https://json.schemastore.org/dependabot-2.0.json',
+              },
+            },
+          }),
+        },
+      },
+    })
+  end,
+}
