+++ .config/nvim/lua/plugins/mkdir.lua
@@ -1,0 +1,4 @@
+return {
+  'jghauser/mkdir.nvim',
+  lazy = true,
+}
