+++ .config/nvim/lua/plugins/onedark.lua
@@ -1,0 +1,12 @@
+return {
+  'navarasu/onedark.nvim', -- https://github.com/navarasu/onedark.nvim
+  lazy = false,
+  priority = 1000,
+  opts = {
+    style = 'dark',
+    ending_tildes = true,
+  },
+  init = function()
+    require('onedark').load()
+  end,
+}
