+++ .config/nvim/stylua.toml
@@ -1,4 +1,4 @@
-column_width = 200
+column_width = 120
 line_endings = "Unix"
 indent_type = "Spaces"
 indent_width = 2
